requirements.txt
python-telegram-bot==20.3
aiohttp==3.9.3